Make default journald config easier to extend
I'm trying to make it easier to add new sources (and new app destinations), once this "system" config is in place.

This commit will only remove fields if we match the `journald` source, and the same with the output. Ideally, we'd do the same for the `journald-format` filter, but I don't know how to do that (besides patching the filter itself). It seems to work for now, so I'll leave it as is.